#38 Game now auto-starts, so 'start game' is no longer required
