+++ ui/players/templates/players/base.html
@@ -41,7 +41,6 @@
                         <li class="active"><a href="{% url 'program' %}">Program<span class="sr-only"> (current)</span></a></li>
                         <li><a href="{% url 'watch' %}">Watch</a></li>
                         <li><a href="{% url 'statistics' %}">Statistics</a></li>
-                        <li><a href="{% url 'start_game' %}">Start Game</a></li>
                     </ul>
                     <ul class="nav navbar-nav navbar-right">
                         {% if request.user.is_authenticated %}
